feature: add builder api
The inspiration for this commit was to challenge myself to create a more
ergonomic api for creating a graph. Now the user can refer to edges by
their names and not their ids as well as have validation on the graph
before you got to create it's parts. Also there is stubbed out logic for
rolling back a graph to run if errors are encountered.